Updates for Idris library changes
In Idris 0.9.19 the semantics of zipWith have changed to allow operating on
lists of different lengths, instead of requiring proofs that lengths are equal.
Thus we no longer need the zipWith' helper function that works the same way as the
new Prelude zipWith.

Also, just after the 0.9.19 release, there have been changes in the Num and Neg typeclasses,
so the (-) operator for Nats is now a separate standalone function, requiring proof that
it wouldn't produce an unrepresentable negative number, while 'minus' retains the usual
behavior of truncating at zero. So we change a (-) to 'minus' to accomodate this change.